requirements.txt
asgiref==3.2.3
boto3==1.14.8
botocore==1.17.8
dj-database-url==0.5.0
Django==3.0.3
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.0.0
psycopg2-binary==2.8.5
pytesseract==0.3.3
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.0
urllib3==1.25.9
whitenoise==5.1.0
django-ckeditor==5.9.0
